% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/please.R
\name{extractMutationTypes}
\alias{extractMutationTypes}
\title{Extract and Validate Mutation Types from VCF Data}
\usage{
extractMutationTypes(vcf_input, ref_genome, context_length)
}
\arguments{
\item{vcf_input}{File path to the VCF file or a CollapsedVCF object.}

\item{ref_genome}{Reference genome object.}

\item{context_length}{Integer specifying the total length of the genomic context,
including the mutated nucleotide.}
}
\value{
A vector of mutation types.
}
\description{
This function extracts and validates mutation types from a VCF file, considering a specified
context length around each mutation site. It ensures that only single nucleotide variants
(SNVs) are considered and reports mutations with reference bases C or T only, avoiding
redundancy by converting mutations with reference bases A or G to their reverse complements.
}
\examples{
vcf<- system.file("extdata", "chr22.vcf.gz", package = "VariantAnnotation")
vcf_data <- readVcf(vcf, genome = "hg19")
vcf_data <- renameSeqlevels(vcf_data, 'chr22')
Hsapiens <- BSgenome.Hsapiens.UCSC.hg19
extractMutationTypes(vcf_data, Hsapiens, 3)

}
